Add cui::panels::splitter namespace
